ambient/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "geocoder>=1.38.1",
    "openai>=2.8.0",
    "openrouter>=0.0.16",
    "pydantic>=2.12.4",
    "python-dotenv>=1.0.0",
    "spotipy>=2.25.1",
]
